# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  helper.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  helper.SuccessResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  paket.CreatePaketRequest:
    properties:
      deskripsi:
        type: string
      harga:
        minimum: 0
        type: integer
      kuota_total:
        minimum: 1
        type: integer
      nama_paket:
        maxLength: 150
        minLength: 3
        type: string
      tanggal_berangkat:
        type: string
    required:
    - harga
    - kuota_total
    - nama_paket
    - tanggal_berangkat
    type: object
  paket.UpdatePaketRequest:
    properties:
      deskripsi:
        type: string
      harga:
        minimum: 0
        type: integer
      is_active:
        type: boolean
      kuota_total:
        minimum: 1
        type: integer
      nama_paket:
        maxLength: 150
        minLength: 3
        type: string
      tanggal_berangkat:
        type: string
    required:
    - harga
    - kuota_total
    - nama_paket
    - tanggal_berangkat
    type: object
  user.ChangePasswordRequest:
    properties:
      new_password:
        maxLength: 100
        minLength: 6
        type: string
      old_password:
        minLength: 6
        type: string
    required:
    - new_password
    - old_password
    type: object
  user.UpdateProfileRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      phone:
        maxLength: 15
        minLength: 10
        type: string
    required:
    - email
    - name
    - phone
    type: object
host: localhost:8080
info:
  contact:
    name: TravelKu
  description: REST API for TravelKu backend (Auth & Users).
  title: TravelKu Backend API
  version: "1.0"
paths:
...
  /pakets:
    get:
      description: Ambil daftar paket wisata, bisa difilter hanya yang aktif
      parameters:
      - description: Filter hanya paket aktif (true/false)
        in: query
        name: active_only
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: List paket wisata
      tags:
      - Pakets
    post:
      consumes:
      - application/json
      description: Buat paket wisata baru dengan kuota dan harga
      parameters:
      - description: Payload pembuatan paket
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/paket.CreatePaketRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tambah paket wisata
      tags:
      - Pakets
  /pakets/{id}:
    delete:
      description: Hapus paket wisata berdasarkan ID
      parameters:
      - description: Paket ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Hapus paket wisata
      tags:
      - Pakets
    get:
      description: Ambil data detail paket wisata berdasarkan ID
      parameters:
      - description: Paket ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Detail paket wisata
      tags:
      - Pakets
    put:
      consumes:
      - application/json
      description: Perbarui informasi paket wisata yang ada
      parameters:
      - description: Paket ID
        in: path
        name: id
        required: true
        type: string
      - description: Payload pembaruan paket
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/paket.UpdatePaketRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Edit paket wisata
      tags:
      - Pakets
  /users:
    get:
      description: Retrieve all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: List users
      tags:
      - Users
  /users/{id}:
    delete:
      description: Delete user by id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - Users
    get:
      description: Retrieve a user by id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update name, email, and phone
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update profile payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Users
  /users/{id}/change-password:
    post:
      consumes:
      - application/json
      description: Change user's password
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Change password payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - Users
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Use: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"